+++ src/components/index.js
@@ -4,3 +4,4 @@ export { Header } from './header/header';
 export { ListView } from './listView/listView';
 export { ListItem } from './listItem/listItem';
 export { Loading } from './loading/loading';
+export { ErrorDisplay } from './errorDisplay/errorDisplay';