+++ src/containers/home/components/searchForm.js
@@ -41,13 +41,6 @@ export default class SearchForm extends React.Component {
 			dispatch(getNearbyVenues(value, radius * 1000));
 		};
 
-		// const selectStyles = {
-		// 	menu: base => ({
-		// 	  ...base,
-		// 	  zIndex: 100
-		// 	})
-		//   };
-
 		return (
 			<div className="form-container">
 				<form>
@@ -57,10 +50,15 @@ export default class SearchForm extends React.Component {
 								<small>Select an Area</small>
 							</label>
 							<Select
+								menuPortalTarget={document.body}
+								styles={{
+									menuPortal: styles => ({
+										...styles,
+										zIndex: 100
+									})
+								}}
 								options={selectFieldDataList}
-								// styles={selectStyles}
 								className="react-select-container mb-3"
-								classNamePrefix="react-select"
 								name="selectedAreaData"
 								onChange={(data, action) => {
 									dispatch(